Return text from tools (#35)
The tools generally return JSON, even when the response is simple text.
It might do something like return a JSON list with one text element. A
bit silly! So here I've rejigged to generally just return plain text.
Should help keep token usage a bit lower. And also makes it easier to
read the responses, which is one of the big features of MCPunk.