Rework file tree (#36)
previously it returns a JSON thingo with the intention of just having to
list out directory names once. But it's a bit silly, a bit annoying to
read. I've just simplified it down to return plain text like below.
Makes response size in chars about half or so the size it was before.

```
dir1: file1.txt
dir2/dir3: file2.txt; file3.txt
...
```