Auto Reload on File Changes (#44)
Closes #38

This is a fairly simple implementation where `watchdog` is used to watch
for changes on the filesystem. Each change is appended to a 'pending
refresh' list. And every 100ms (by default) the code refreshes them in
bulk. This 100ms interval allows things to accumulate, thus allowing for
sensible multiprocessing. And it avoids things like where IDEs create
temporary files (pycharm seems to do this on save).

It's quite possibly a bit racy. I haven't intended to make it bullet
proof, and can always do a second pass on it.